backend/requirements.txt
flask==2.0.1
werkzeug==2.0.3
openai==1.0.0
spotipy==2.23.0
python-dotenv>=0.21.0
requests>=2.31.0
flask-cors==3.0.10
urllib3>=2.2.2 